Fixed repeat of "/language list"

+++ Console Edition/src/Utility/Commands/LanguageCommand.vb
@@ -17,50 +17,47 @@
 
         If command = "language" Then
             Console.ForegroundColor = ConsoleColor.Yellow
-            Console.WriteLine("Correct command is /language <languagename>")
+            Console.WriteLine("Correct command is /language <languagename|list>")
 
-        ElseIf command = "language list" Then
+        ElseIf command = "language " + args(1) Then
+            If args(1) = "italian" Or args(1) = "16" Then
+                language = 16
+                Console.ForegroundColor = ConsoleColor.Blue
+                Console.WriteLine("Languange set to Italian")
 
-            Console.WriteLine("1) Afrikaans" + vbTab + vbTab + vbTab + "16) Italian")
-            Console.WriteLine("2) Arabic" + vbTab + vbTab + vbTab + "17) Japanese")
-            Console.WriteLine("3) Catalan" + vbTab + vbTab + vbTab + "18) Korean")
-            Console.WriteLine("4) Chinese Simplified" + vbTab + vbTab + "19) Norwegian")
-            Console.WriteLine("5) Chinese Traditional" + vbTab + vbTab + "20) Polish")
-            Console.WriteLine("6) Czech" + vbTab + vbTab + vbTab + "21) Portuguese")
-            Console.WriteLine("7) Danish" + vbTab + vbTab + vbTab + "22) Portuguese, Brazilian")
-            Console.WriteLine("8) Dutch" + vbTab + vbTab + vbTab + "23) Romanian")
-            Console.WriteLine("9) English" + vbTab + vbTab + vbTab + "24) Russian")
-            Console.WriteLine("10) Finnish" + vbTab + vbTab + vbTab + "25) Serbian(Cyrillic)")
-            Console.WriteLine("11) French" + vbTab + vbTab + vbTab + "26) Spanish")
-            Console.WriteLine("12) German" + vbTab + vbTab + vbTab + "27) Swedish")
-            Console.WriteLine("13) Greek" + vbTab + vbTab + vbTab + "28) Turkish")
-            Console.WriteLine("14) Hebrew" + vbTab + vbTab + vbTab + "29) Ukrainian")
-            Console.WriteLine("15) Hungarian" + vbTab + vbTab + vbTab + "30) Vietnamese")
-        Else
-            If command.ToLower = "language " + args(1) Then
-                If args(1) = "italian" Or args(1) = "16" Then
-                    language = 16
-                    Console.ForegroundColor = ConsoleColor.Blue
-                    Console.WriteLine("Languange set to Italian")
+            ElseIf args(1) = "english" Or args(1) = "9" Then
+                language = 9
+                Console.ForegroundColor = ConsoleColor.Blue
+                Console.WriteLine("Languange set to English")
 
-                ElseIf args(1) = "english" Or args(1) = "9" Then
-                    language = 9
-                    Console.ForegroundColor = ConsoleColor.Blue
-                    Console.WriteLine("Languange set to English")
+            ElseIf args(1) = "list" Then
+                Console.WriteLine("1) Afrikaans" + vbTab + vbTab + vbTab + "16) Italian")
+                Console.WriteLine("2) Arabic" + vbTab + vbTab + vbTab + "17) Japanese")
+                Console.WriteLine("3) Catalan" + vbTab + vbTab + vbTab + "18) Korean")
+                Console.WriteLine("4) Chinese Simplified" + vbTab + vbTab + "19) Norwegian")
+                Console.WriteLine("5) Chinese Traditional" + vbTab + vbTab + "20) Polish")
+                Console.WriteLine("6) Czech" + vbTab + vbTab + vbTab + "21) Portuguese")
+                Console.WriteLine("7) Danish" + vbTab + vbTab + vbTab + "22) Portuguese, Brazilian")
+                Console.WriteLine("8) Dutch" + vbTab + vbTab + vbTab + "23) Romanian")
+                Console.WriteLine("9) English" + vbTab + vbTab + vbTab + "24) Russian")
+                Console.WriteLine("10) Finnish" + vbTab + vbTab + vbTab + "25) Serbian(Cyrillic)")
+                Console.WriteLine("11) French" + vbTab + vbTab + vbTab + "26) Spanish")
+                Console.WriteLine("12) German" + vbTab + vbTab + vbTab + "27) Swedish")
+                Console.WriteLine("13) Greek" + vbTab + vbTab + vbTab + "28) Turkish")
+                Console.WriteLine("14) Hebrew" + vbTab + vbTab + vbTab + "29) Ukrainian")
+                Console.WriteLine("15) Hungarian" + vbTab + vbTab + vbTab + "30) Vietnamese")
+            End If
 
-                End If
+            If checklanguage And args(1) <> "list" Then '->args(1) <> "list"<- because prevent an inutil check :P
+                My.Computer.FileSystem.DeleteFile("C:\Program Files\PocketMine-ManagerServers\Data\langselection.pm")
+                My.Computer.FileSystem.WriteAllText("C:\Program Files\PocketMine-ManagerServers\Data\langselection.pm", language, True)
 
-                If checklanguage Then
-                    My.Computer.FileSystem.DeleteFile("C:\Program Files\PocketMine-ManagerServers\Data\langselection.pm")
-                    My.Computer.FileSystem.WriteAllText("C:\Program Files\PocketMine-ManagerServers\Data\langselection.pm", language, True)
-
-                Else
-                    Console.ForegroundColor = ConsoleColor.Red
-                    Console.WriteLine("An error occured")
-
-                End If
+            ElseIf checklanguage = False Then
+                Console.ForegroundColor = ConsoleColor.Red
+                Console.WriteLine("An error occured")
 
             End If
+
         End If
     End Sub
